refactoring report templates
