+++ safe/impact_reports/test/test_road_exposure_report_mixin.py
@@ -50,6 +50,7 @@ class RoadExposureReportMixinTest(unittest.TestCase):
                 'Bike': 1.2})
         ])
         self.road_mixin.affected_road_categories = ['Flooded']
+        self.maxDiff = None
 
     def tearDown(self):
         """Run after each test."""
@@ -75,7 +76,7 @@ class RoadExposureReportMixinTest(unittest.TestCase):
             'exposure': 'road',
             'impact summary': {
                 'attributes': ['category', 'value'],
-                'fields': [['Unaffected', 0], ['Total', 0]]},
+                'fields': [['Unaffected (m)', 0], ['Total (m)', 0]]},
             'impact table': {
                 'attributes': [
                     'Road Type', 'Unaffected', 'Total'],
@@ -90,9 +91,9 @@ class RoadExposureReportMixinTest(unittest.TestCase):
         roads_breakdown = self.road_mixin.impact_table()['fields']
 
         expected = [
-            ['Main', 2, 131.3, 133.3],
-            ['Side', 5.5, 4.5, 10],
-            ['Bike', 1.2, 0.0, 1.2]
+            ['Main (m)', 2, 131.3, 133.3],
+            ['Side (m)', 5.5, 4.5, 10],
+            ['Bike (m)', 1.2, 0.0, 1.2]
         ]
 
         self.assertEquals(roads_breakdown, expected)
@@ -128,9 +129,9 @@ class RoadExposureReportMixinTest(unittest.TestCase):
             'impact summary': {
                 'attributes': ['category', 'value'],
                 'fields': [
-                    ['Flooded', 8.7],
-                    ['Unaffected', 135.8],
-                    ['Total', 144.5]
+                    ['Flooded (m)', 8.7],
+                    ['Unaffected (m)', 135.8],
+                    ['Total (m)', 144.5]
                 ]
             },
             'impact table': {
@@ -138,9 +139,9 @@ class RoadExposureReportMixinTest(unittest.TestCase):
                     'Road Type', 'Flooded', 'Unaffected', 'Total'
                 ],
                 'fields': [
-                    ['Main', 2, 131.3, 133.3],
-                    ['Side', 5.5, 4.5, 10],
-                    ['Bike', 1.2, 0.0, 1.2]
+                    ['Main (m)', 2, 131.3, 133.3],
+                    ['Side (m)', 5.5, 4.5, 10],
+                    ['Bike (m)', 1.2, 0.0, 1.2]
                 ]
             },
             'notes': {
