+++ safe/impact_template/test/test_generic_report_template.py
@@ -9,19 +9,19 @@ InaSAFE Disaster risk assessment tool developed by AusAid and World Bank
      (at your option) any later version.
 
 """
-__author__ = 'ismailsunni'
-__project_name__ = 'inasafe-dev'
-__filename__ = 'test_road_report_template'
-__date__ = '4/28/16'
-__copyright__ = 'imajimatika@gmail.com'
-
 import unittest
 import os
 import json
 from safe.common.utilities import temp_dir
 
-from safe.impact_template.road_building_report_template import (
-    RoadBuildingReportTemplate)
+from safe.impact_template.generic_report_template import (
+    GenericReportTemplate)
+
+__author__ = 'ismailsunni'
+__project_name__ = 'inasafe-dev'
+__filename__ = 'test_generic_report_template'
+__date__ = '4/28/16'
+__copyright__ = 'imajimatika@gmail.com'
 
 TEMP_DIR = temp_dir(sub_dir='test/template')
 TEST_DIR = os.path.dirname(__file__)
@@ -30,15 +30,14 @@ ROAD_JSON_FILE = os.path.join(TEST_DIR, 'data', 'road_impact.json')
 BUILDING_JSON_FILE = os.path.join(TEST_DIR, 'data', 'building_impact.json')
 
 
-class TestRoadReportTemplate(unittest.TestCase):
+class TestGenericReportTemplate(unittest.TestCase):
 
     def test_road_report_template(self):
-        """Test Road Report Template
-        """
+        """Test Road Report Template."""
         with open(ROAD_JSON_FILE) as json_file:
             impact_data = json.load(json_file)
 
-        road_report_template = RoadBuildingReportTemplate(
+        road_report_template = GenericReportTemplate(
             impact_data=impact_data)
         report = road_report_template.generate_message_report()
         self.assertIn(
@@ -46,18 +45,17 @@ class TestRoadReportTemplate(unittest.TestCase):
 
         expected = '''
 ---
-**17459.7365945 (m)**, 41,116, 58,576---
+**17459.7365945**, 41,116, 58,576---
 '''
         result = report.message[1].to_text()
         self.assertIn(expected, result)
 
     def test_building_report_template(self):
-        """Test Building Report Template
-        """
+        """Test Building Report Template."""
         with open(BUILDING_JSON_FILE) as json_file:
             impact_data = json.load(json_file)
 
-        building_report_template = RoadBuildingReportTemplate(
+        building_report_template = GenericReportTemplate(
             impact_data=impact_data)
         report = building_report_template.generate_message_report()
         self.assertIn(
